requirements.txt
streamlit
requests
python-dotenv
reportlab
pypdf2
sqlalchemy